database/pyproject.toml
[project]
name = "database"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "dotenv>=0.9.9",
    "mysql-connector-python>=9.4.0",
]
